requirements.txt
certifi==2020.12.5
chardet==4.0.0
emoji==1.2.0
idna==2.10
joblib==1.0.1
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
pymongo==3.11.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tweepy==3.10.0
urllib3==1.26.3
wget==3.2
yapf==0.30.0
